Add account pages

+++ src/formatDate.ts
@@ -1,4 +1,4 @@
-const dateFormatter = new Intl.DateTimeFormat('en-US', { year: 'numeric', month: 'long', day: 'numeric' });
+const dateFormatter = new Intl.DateTimeFormat('en-US', { year: 'numeric', month: 'short', day: 'numeric' });
 
 export const formatDate = (d: Date): string => {
   return dateFormatter.format(d);
